Add option to set lunchtime
If we tried to pass just the entrance time before, we were unable to
know the exact time we should end the day.

This commit addresses this issue, and now the user is able to pass only
the entrance hour and let the application tells the exact time he/she
needs to finish the day.

A new option (-l or --lunch) was added, and defaults to "1" (1h lunchtime).

Half time should passed as "0.5".