Add Things integration
I thought in way to remind me when I need to clock out. As I use Things
as my main task manager, I added this integration using the URL schema
from it, as show in its docs.

I had to use a NPM library call `open` (formerly `opn`) to open the
schema and add the integration. Good news, it's working.

TODO: Add an option to get rid of the message and change the actual text
of the reminder.